Ensure RPs validate the `alg` header on the ID Token
The current text opens RPs up to a potential attack, where an attacker
can use the same key but a different algorithm.  The worst version of
this is reusing an RS256 public key as an HS256 symmetric key.

The solution is to ensure that the `alg` parameter is validated before
the JWT signature itself is verified.  There are two methods we could
take:

 1. get the RP to only accept an allowlist of algorithms (which would
    include `ES256`) (this is simple but not very future-proof)
 2. ensure the RP matches the `alg` with the `alg` specified at the
    jwks_uri: https://oidc.account.gov.uk/.well-known/jwks.json

This commit takes the second approach.