Update github references to reflect this repos new org
We've moved this repository from alphagov (Cabinet Office's GitHub) to
govuk-one-login (Our Programme's GitHub). Some references were unchanged
leading the HTML proofer to be upset.

Note there are still some valid alphagov links left.
We do not appear to have moved the [di-auth-stub-relying-party][1]
Nor the code of conduct[2]. Those will need to be updated if/when we
move those across.

[1]: https://github.com/alphagov/di-auth-stub-relying-party
[2]: https://github.com/alphagov/code-of-conduct