Remove .secrets.baseline
This file was generated as part of the initial commit and hasn't been
touched since.

It's a config file for [detect-secrets][1] which is a pre-commit hook
scanning tool looking for stuff that looks a bit like secrets and
catching them before you commit and send them to github repos.

Two reasons we're removing this:

1. Detect Secrets and Pre-commit hooks do not appear to be widely used
   by this team, nor integrated into our documentation. If we think
   there's value in this approach we should look at bringing it back
   alongside a wider bit of team training, and with a view to our
   README.md and onboarding.

2. We can effectively deem this repository as a low risk for secrets
   leakage, whilst it does have a deployment element and there is always
   a risk there, most of the changes are content related. Those with
   access to keys that could be leaked are small in number. Were we to
   spot any leakage, it's those same folks who could quickly rotate
   keys. And this is not an application that contains PII. Worst case we
   can blow it away and redeploy as a static site.

That's not to preclude us looking at a better security threat assessment
here, just to say that for now it's not a burning priority here and if
we change our mind we should bring it back as part of a fuller
implimentation.

[1]: https://github.com/Yelp/detect-secrets