Downgrade to Ruby 3.2.2 due to Apple processer issues
We're getting some C library errors when we try to build Ruby 3.3.0 on
apple processors. That's a problem becuase quite a few of our Tech
Writers are using them.

The issue is known to Ruby here[1], and they'll merge a fix as Ruby
v3.3.1 reasonably soon. We can upgrade then.

However until then ./preview-with-docker.sh is broken. This should
resolve the issue.

[1]: https://github.com/ruby/ruby/pull/9371