Relaxes codeowners
Current approach wasn't working the conditionality was broken so we
couldn't get the targeting we wanted where Devs / Architects own the
platform and tech writers own the content.

Instead I'm relaxing the technical constraints and moving this to a
policy based approach.

We'll trust people involved in the pod or at the right program level to
make sensisble descisions, but come to an understanding between
engineering, architecture and tech writers about who should be involved
in what kind of change.

We'll also ensure that Tech writers can have a periodic review point to
ensure that any changes made over time are not leading to declining
content standards.