Use a quote character in partial links that cannot be confused for code
There's a difference between the characters:
’ and '

One is meaningful in content, and the other is meaningful in code, I
think we should take care to use the content one within content, and
should probably try and lint for it where possible.

My hypothesis to test here is that HTML proofer was getting confused,
either directly or because something was going wrong during the markdown
to HTML compilation. Going to try pushing this up now and seeing.

HTML proofer seems happy locally!