feature(decap-cms): Add /editor page that loads Decap CMS
This is an attempt to solve a general issue we have on the program.
We believe in managing docs in markdown [1]. Our core tech doc stack has
a good deal of GDS value encoded in it, broadly it allows us to spin up
simple docs sites quickly.

We've done that here in this repo, and also in other areas of the
program.

However this creates an issue, the presumption of this approach is that
everyone who has a stake in these docs (authors, editors, reviewers),
would be people comfortable with a command line and GitHub as an
interface.

I think that is ceasing to be true. There's now a wider range of
stakeholders who have an interest in our docs.

I've tried experiments with other CMS tools, both SASS products and
products we can attach to our tech docs stack. What we see here is the
best I've come across so far.

It allows us to keep the existing approach intact, allowing developers
to continue to write docs in line with lightweight agile principles.

However it opens that process up to folks who are not comfortable with
(or simply prefer), a non CLI based approach. Through widening
participation my hope is we can solve a need that is otherwise driving a
bit of fragmentation over where people write docs, with Devs using
document repositories and others seeking WYSIWYG enabled tools.

Decap tries to solve that user need.

[1]: https://technology.blog.gov.uk/2017/08/25/why-we-use-a-docs-as-code-approach-for-technical-documentation/