Add PathBuilder helper
The theory here is that we have a range of conditionality that we need to test for to ensure the
asset paths build correctly.

I think it's a bit easier if we abstract this out of the view layer into a helper.
I'm not actually writing tests here but we could test this indepenetly if we wanted.

The development? helper only works in the view layer in middleman so it's best to make this a functional
style method call that just takes in the app class and calls development and config off it.

The outcome is the same, but we can also use early returns to knock out endpoints we want is a way folks
might find a little easier to reason about