Add Brand helper
Some of the conditionality seems to key off the yaml, I wonder if it's
easier to make that a more readable and explicit call?

Not sure if "blue" is the right way of wording it but Brand.rebrand? felt
a bit odd, will leave that up to orch to decide on best domain language for?